Refactored form rendering
A lot of dead code has been removed. Also a lot of "old" code has been simplified, so it can be removed.

The rendering of form controls is alot more direct now. So the code is a lot easier to follow. Also
where possible hsx has been used instead of the "old" blaze syntax.