Fixed code generators not finding the right table if word is not english
In case we're dealing with e.g. german words in the code gen (e.g. a table called immobilien instead of properties), the pluralization and singularization fails and thus the table is not found in the Schema.sql. This leads to the code gen not generating the form fields.

Fixed this problem by falling back to the user input if the pluralized/singularized value has no matching table in Schema.sql