Optimized makefile to speed up compile times of jobs and scripts
Previously when compiling a script or job GHC was not reusing any of the previously compiled object files. So for every script and every job a lot of files would be recompiled that might have already been compiled when making the normal server binary. This made production builds much slower than they need to be