Fixed live reloading not working if the dev server is not running on localhost
This improves support for GitPod, GitHub Codespaces and setups where the dev server is running on a different host.

The livereload websocket server has been merged into the toolserver. The server previously running at :8002 is now running as part of the IDE server at :8001.

The livereload.js script tag has a new data-ws attribute to pass the websocket url to the script. For supporting existing apps this will by default use localhost:8001 when no value is defined. All newly generated apps will have this defined as part of their boilerplate. If the IHP_IDE_BASEURL is defined, the data-ws attribute will be set to that