Allow configuring the client session key via IHP_SESSION_SECRET env var
When the IHP_SESSION_SECRET env var is set, it's used instead of the Config/client_session_key.aes file.

When the Config/client_session_key.aes is missing an no IHP_SESSION_SECRET is defined, a random key will be generated and temporarly be used. This means that all sessions (e.g. login sessions) will be gone after the server is restarted and need to re-login.