Fixed fill not able to deal with certain JSON bodies
The previous implementation of `fill` only worked with JSON bodies of the following form:

{"key": "value"}

It failed for requests where the value was not encoded as a string, such as these:

{"boolField": true}
{"intField": 1337}

This was caused by the implementation of fill using `paramOrNothing` to check for the existence of a parameter in the request. This has worked fine for url encoded request bodies as there the input has no type and can always be parsed as a string (e.g. `1337` can always be parsed as a string, while in json requests we have type information and only accept `"string"` as a string and raising an error for `1337` because it's a number)