Added auto-migration based on operation tracking approach
Every time a table, we save the CREATE TABLE statement to Application/Migration/unmigrated-changes.sql. Same for new columns and other changes in the future.

When a new migration is generated, we take all the SQL statements in Application/Migration/unmigrated-changes.sql as the starting point for the new migration.