Improved database query performance if RLS is enabled
Previously when RLS was enabled it was firing multiple queries just to set up the context, before running the actual query. For DataSync we've already been using a more efficient way of just firing a single query. This commit applies the same approach to the general query interface as well