Fixes qualified identifiers in policies causing a diff in the migrations
A policy in the form `CREATE POLICY a ON t USING t.f = public.f` contains a qualified sub-expression `t.f`. When dumping this policy from pg_dump we get a unqualified `f = public.f` condition instead. So we need to normalize the Policy to the normal form `CREATE POLICY a ON t USING f = public.f`