DataSync. fixed type error in recordMatchesQuery
in the branch for LiteralExpression the expression was not correctly unpacked. This resulted in evaluateDynamicValue returning undefined. This resulted in recordMatchesQuery sometimes returning true when it shouldn't.

In practise this means that on a createRecord call, sometimes a record will be added to a useQuery result optimistically, even when the query would not match that record.