ihp-hsx: Split out core library that doesn't depend on blaze
If the HSX syntax can be ported to more HTML libraries than just
blaze-html, it has a better chance of being more widely adopted across
the haskell ecosystem. Split out ihp-hsx into two parts. A core library
that only parses, and doesn't depend on blaze. Second: the blaze
dependent part with template haskell to generate blaze-specific code.

The blaze library continues to retain the name ihp-hsx, and re-exports
all of the modules from ihp-hsx-core. It should be completely
transparent to existing users of the library.