ihp-hsx: Add Lucid2 port
Add a Lucid2 backend for HSX (the language) using ihp-hsx-core. This
has no dependencies on blaze-html or blaze-markup. This does two things:
It makes it easier to use ihp-hsx on a project using lucid, and it draws
a roadmap to others that might want to put ihp-hsx-core to use in a
different way. By having a 2nd example to look at, it's easier to see
what was blaze specific, and what is inherent to HSX.